FIX - Смена типа ID-шников в ParentAndChild на long в местах, которые сперва не попались на глаза.

+++ UserService/src/test/java/modelconnectors/ParentsDatabaseConnectorTest.java
@@ -14,7 +14,7 @@ class ParentsDatabaseConnectorTest {
 
     @Test
     void getChildrenIDsOfUser() throws SQLException, DataBaseConnectionException {
-        List<Integer> idOfChildren = pdc.getChildrenIDsOfUser(1);
+        List<Long> idOfChildren = pdc.getChildrenIDsOfUser(1);
 
         assertNotEquals(0 , idOfChildren.size());
         System.out.println(idOfChildren);
@@ -22,7 +22,7 @@ class ParentsDatabaseConnectorTest {
 
     @Test
     void getParentsIDsOfUser() throws SQLException, DataBaseConnectionException {
-        List<Integer> idOfParents = pdc.getParentsIDsOfUser(2);
+        List<Long> idOfParents = pdc.getParentsIDsOfUser(2);
 
         assertNotEquals(0 , idOfParents.size());
         System.out.println(idOfParents);
@@ -32,12 +32,12 @@ class ParentsDatabaseConnectorTest {
     void compareChildrenAndParentsTest() throws SQLException, DataBaseConnectionException {
         int testParentId = 1;
 
-        List<Integer> idOfChildren = pdc.getChildrenIDsOfUser(testParentId);
+        List<Long> idOfChildren = pdc.getChildrenIDsOfUser(testParentId);
         System.out.println("Children of user {id=" + testParentId + "}:\n" + idOfChildren);
 
         if (!idOfChildren.isEmpty()){
-            int firstChildId = idOfChildren.get(0);
-            List<Integer> idOfParents = pdc.getParentsIDsOfUser(firstChildId);
+            long firstChildId = idOfChildren.get(0);
+            List<Long> idOfParents = pdc.getParentsIDsOfUser(firstChildId);
             System.out.println("Parents of child {id=" + firstChildId + "}:\n" + idOfParents + "\nShould contain '" + testParentId + "'!");
             assertTrue(idOfParents.contains(testParentId));
         }
